requirements.txt
django==1.9.13
django-crontab==0.7.1
django-filter==1.1.0
djangorestframework==3.3.2
dominate==2.2.1
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
Markdown==2.6.11
MarkupSafe==0.23
MySQL-python==1.2.3


